nanobot-webchat/pyproject.toml
[project]
name = "nanobot-webchat"
version = "1.0.0"
description = "WebSocket channel plugin for nanobot-ai — enables web clients to chat with the agent"
requires-python = "==3.14.*"
dependencies = [
    "nanobot-ai",
    "websockets>=15.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
nanobot-ai = { url = "https://github.com/HKUDS/nanobot/archive/e7d371ec1e6531b28898ec2c869ef338e8dd46ec.zip" }

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["nanobot_webchat*"]

[project.entry-points."nanobot.channels"]
webchat = "nanobot_webchat:WebChatChannel"
